.history/requirements_20250622024444.txt
google-adk
<<<<<<< HEAD
yfinance==0.2.56
=======
>>>>>>> 3044804c33d248db204947d1dd76e60914688b8e
psutil==5.9.5
litellm==1.66.3
google-generativeai
python-dotenv==1.1.0
deprecated
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
